+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Test against more openHAB versions
 - New Thing struct
+- New Channel struct
 
 ## [0.3.2] - 2025-04-25